Expose golden test data
